Fix "Es ist ein unerwarteter Fehler aufgetreten"
This fixes the error "Es ist ein unerwarteter Fehler aufgetreten" when checking for an appointment with a Vermittlungscode.
This error appeared consistenly for me, because the bot sends a request immediately when the page is loaded. With this commit, it sleeps 2 seconds each after loading the site putting in the Vermittlungscode, this permanently fixed the error for me.